# internal/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: Via Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for the Via backend (gateway server).

    Notes:
    - Async command endpoints return `202 Accepted` with an operation id.
    - Authentication supports either bearer JWT (`Authorization`) or API key (`X-API-Key`).
    - WebSocket routes are documented as HTTP GET upgrade endpoints.
servers:
  - url: http://localhost:9090
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/gps/update:
    post:
      tags: [Realtime]
      summary: Publish GPS update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GPSUpdate'
      responses:
        '202':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishAck'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: Upstream broker error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/trip/start:
    post:
      tags: [Realtime]
      summary: Publish trip event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealtimeEvent'
      responses:
        '202':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishAck'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/events/publish:
    post:
      tags: [Realtime]
      summary: Publish realtime event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealtimeEvent'
      responses:
        '202':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishAck'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /ws:
    get:
      tags: [Realtime]
      summary: Realtime WebSocket stream
      description: |
        Upgrades to WebSocket and streams topic data.
        Use `topic=operations&operation_id=<id>` to stream operation status.
      parameters:
        - $ref: '#/components/parameters/FleetIDQuery'
        - $ref: '#/components/parameters/VehicleIDQuery'
        - $ref: '#/components/parameters/TopicQuery'
        - $ref: '#/components/parameters/OperationIDQuery'
      responses:
        '101':
          description: Switching Protocols
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  parameters:
    FleetIDQuery:
      name: fleet_id
      in: query
      required: false
      schema:
        type: string
...
    VehicleIDQuery:
      name: vehicle_id
      in: query
      required: false
      schema:
        type: string
...
    TopicQuery:
      name: topic
      in: query
      required: false
      schema:
        type: string
        enum: [gps, trip, ops, events, operations]
        default: gps
    OperationIDQuery:
      name: operation_id
      in: query
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        public_message:
          type: string
...
    PublishAck:
      type: object
      required: [status, subject]
      properties:
        status:
          type: string
        subject:
          type: string
    GPSUpdate:
      type: object
      required: [fleet_id, vehicle_id, latitude, longitude]
      properties:
        fleet_id:
          type: string
        vehicle_id:
          type: string
        driver_id:
          type: string
        route_id:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        speed_kph:
          type: number
          format: double
        heading:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
    RealtimeEvent:
      type: object
      required: [fleet_id, vehicle_id]
      properties:
        fleet_id:
          type: string
        vehicle_id:
          type: string
        driver_id:
          type: string
        route_id:
          type: string
        topic:
          type: string
        message:
          type: string
        event:
          type: string
        timestamp:
          type: string
          format: date-time
...